+++ CHANGELOG.md
@@ -15,10 +15,16 @@
 
 ## Unreleased
 
+- Added the “Parse embeds” field setting. ([#409](https://github.com/craftcms/ckeditor/pull/409))
+
+## 4.8.0 - 2025-04-30
+
 - Added the “Field Limit” field setting, which can be set to a character limit or word limit. ([#384](https://github.com/craftcms/ckeditor/pull/384))
 - Added the “GraphQL Mode” field setting. ([#403](https://github.com/craftcms/ckeditor/pull/403))
+- Editor toolbars now have dedicated nested entry creation buttons, for entry types that have icons. Any entry types that don’t have an icon are now grouped into a “+” menu. ([#404](https://github.com/craftcms/ckeditor/pull/404))
 - Improved the styling of CKEditor fields. ([craftcms/cms#17164](https://github.com/craftcms/cms/discussions/17164))
 - The “Anchors” CKEditor plugin has been replaced with CKEditor’s new built-in [Bookmarks](https://ckeditor.com/docs/ckeditor5/latest/features/bookmarks.html) plugin. ([#397](https://github.com/craftcms/ckeditor/pull/397))
+- Removed the “‘New’ Button Label” field setting. ([#404](https://github.com/craftcms/ckeditor/pull/404))
 - Updated to CKEditor 5 45.0.0. ([#397](https://github.com/craftcms/ckeditor/pull/397))
 - Added `craft\ckeditor\Field::$characterLimit`.
 - Added `craft\ckeditor\Field::$fullGraphqlData`.
@@ -29,6 +35,7 @@
 - Added `craft\ckeditor\gql\CkeditorMarkup`.
 - Added `craft\ckeditor\gql\Generator`.
 - Fixed a bug where it wasn’t easily possible to drag a toolbar button to the last position in the toolbar, if that meant wrapping the buttons to a new row.
+- Fixed a JavaScript error that could prevent the toolbar configurator from loading, when editing a CKEditor config within a slideout. ([#401](https://github.com/craftcms/ckeditor/pull/401))
 
 ## 4.7.0 - 2025-04-21